% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peace_books_ner_sentiment.R
\docType{data}
\name{peace_books_ner_sentiment}
\alias{peace_books_ner_sentiment}
\title{Peace Books Corpus with POS and Sentiment Tagged Tokens}
\format{
A data frame with 107013 rows and 7 variables.
\describe{
\item{doc_id}{document identifier. The identifier consists of HathiTrust ID of the volume and the Peace Book volume number}
\item{token}{Token from the text body of the Peace Book volume, ie word.}
\item{lemma}{Base form of the token, with no inflectional suffixes.}
\item{pos}{Coarse-grained part-of-speech from the Universal POS tag set. }
\item{tag}{Fine-grained part-of-speech..}
\item{entity}{Name Entity Recognition tag.}
\item{sentiment}{Sentiment polarity value assigned by the Bing dictionary in textdata package.}
}
}
\source{
Extracted, edited, and transformed into a dataset by Stanislav Pejša.
}
\usage{
peace_books_ner_sentiment
}
\description{
This dataset contains annotated tokens from the Peace Book volumes created by
the Historical Section of the British Foreign Office and published in 1920.
This dataset contains only those tokens that were included in the Bing
sentiment dictionary that is part of the `textdata` package.
The POS and entity tagging was preserved only for the tokens with sentiment
value, tokens without sentiment values were removed. This means that
this dataset cannot be used for analysing NER or POS annotations, because
not all annotated tokens in the corpus are included in this dataset.
The POS and NER annotation were done by using the package `spacyr`.
}
\examples{
  peace_books_ner_sentiment
}
\keyword{datasets}
