% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peace_books_pages.R
\docType{data}
\name{peace_books_pages}
\alias{peace_books_pages}
\title{Peace Book Page by Page}
\format{
A data frame with 370112 rows and 6 columns.
\describe{
\item{doc_id}{document identifier consists of HathiTrust ID of the band and the volume number}
\item{text}{Contains the text body on  a page.}
\item{page}{Page on which the text occured.}
\item{volume_title}{Title of the PB volume.}
\item{volume_number}{Volume number of the Peace Book volume.}
\item{part}{Section in the volume in which the text occured.}
}
}
\source{
Created by Stanislav Pejša
}
\usage{
peace_books_pages
}
\description{
This dataset contains list of all referenced works published
in the Peace Books in 1920.
}
\examples{
  peace_books_pages
}
\keyword{datasets}
