% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaceBooksR-package.R
\docType{package}
\name{peaceBooksR-package}
\alias{peaceBooksR}
\alias{peaceBooksR-package}
\title{peaceBooksR: Peace Books Prepared For the Paris Peace Conference}
\description{
Description: This package contains text of the Peace Books prepared by the Historical Section of the British Foreign Office and published in 1920. The package contains the texts of each volumes and and then the files that contain all texts in in the file. In the dataset peace_books_pages each line of text correspond to to a page, in the dataset peace_books_volumes, each line corresponds to the content of the whole volume. The package also contains dataset peace_books_authorities that contains data extracted from the list of references from each volume with the key bibliographic metadata. The dataset peace_books_sentiment contains tokens that were matched against the `Bing` sentiment dictionary. These tokens are associated with a sentiment value, as well as their lemmas, POS tags, and if available with the name entity recognition (NER) value. The corpora contains only the body of texts, section titles, footnotes, running headers, tables, indices, and list of references (authorities) were removed.
}
\author{
\strong{Maintainer}: Stanislav Pejša \email{spejsa@purdue.edu} (\href{https://orcid.org/0000-0001-5057-262X}{ORCID})

}
\keyword{internal}
