% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peace_books_pages.R
\docType{data}
\name{peace_books_pages}
\alias{peace_books_pages}
\title{Dataset Peace Book Corpus by Pages}
\format{
A data frame with 370112 rows and 6 variables.
\describe{
\item{doc_id}{document identifier. The identifier consists of HathiTrust ID of the volume and the Peace Book volume number}
\item{text}{Contains the text body on a page.}
\item{page}{Page on which the text occured.}
\item{volume_title}{Title of the Peace Book volume.}
\item{volume_number}{Volume number of the Peace Book volume.}
\item{part}{Section in the volume in which the text occured.}
}
}
\source{
Extracted, edited, and transformed into a dataset by Stanislav Pejša.
}
\usage{
peace_books_pages
}
\description{
This dataset contains the text body of the Peace Book volumes created by
the Historical Section of the British Foreign Office and published in 1920.
Some volumes were not published and volumes 155 through 157 were not
included in the corpus. This datasets preservers the structure of
the published corpus where each row correspond to a page of a volume.
}
\examples{
  peace_books_pages
}
\keyword{datasets}
