% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peace_books_volumes.R
\docType{data}
\name{peace_books_volumes}
\alias{peace_books_volumes}
\title{Peace Books by Volume}
\format{
A data frame with 160 rows and 5 columns.
\describe{
\item{doc_id}{document identifier consists of HathiTrust ID of the band and the volume number}
\item{text}{Contains the complete text body of the volume.}
\item{volume_title}{Title of the PB volume.}
\item{volume_number}{Volume number of the Peace Book volume.}
\item{continent}{Region to which was the referenced work related.}
}
}
\source{
Created by Stanislav Pejša
}
\usage{
peace_books_volumes
}
\description{
This dataset contains list of all referenced works published
in the Peace Books in 1920.
}
\examples{
  peace_books_volumes
}
\keyword{datasets}
