% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaceBooksR-package.R
\docType{package}
\name{peaceBooksR-package}
\alias{peaceBooksR}
\alias{peaceBooksR-package}
\title{peaceBooksR: Peace Books Prepared For the Paris Peace Conference}
\description{
This package contains text of the Peace Books prepared by the Historical Section of the Foreign Office and published in 1920. The package contains the texts of each volumes and and then the files that contain all texts in ine file. In the dataset peace_books_pages each line of text correspond to to a page, in the dataset peace_books_volumes, each line corresponds to the content of the whole volume. The copora contains only the body of texts, section titles, footnotes, tunning headers, tables, indices, and list of references (authorities) were removed.
}
\keyword{internal}
